+++ lending/src/sdk/models/shared/invoicelineitem.ts
@@ -112,8 +112,6 @@ export class InvoiceLineItem extends SpeakeasyBase {
 
     /**
      * Reference to the tracking categories to which the line item is linked.
-     *
-     * @deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
      */
     @SpeakeasyMetadata({ elemType: TrackingCategoryRef })
     @Expose({ name: "trackingCategoryRefs" })