+++ lending/src/sdk/models/shared/items.ts
@@ -7,16 +7,9 @@ import { Expose } from "class-transformer";
 
 export class Items extends SpeakeasyBase {
     /**
-     * Amount of tax withheld.
+     * Unique identifier for the group.
      */
     @SpeakeasyMetadata()
-    @Expose({ name: "amount" })
-    amount: number;
-
-    /**
-     * Name assigned to withheld tax.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "name" })
-    name: string;
+    @Expose({ name: "id" })
+    id?: string;
 }